single-player random art generator
A pygame with websockets that has a server and multiple clients that passes a
game object over a network. Currently supports just one player and does not keep
track of scores, gamemaster status, or additional menu items. The game also only
works on the localhost, although this can probably be changed to a known IP
address for communication over the same network.

Future work will include allowing local network communication, multiple users,
adding more artworks and artwork information, and additional menus that allow a
gamemaster to start and stop the game as they see fit.